+++ packages/create-plugin/src/constants.ts
@@ -1,5 +1,5 @@
-export const CLI_VERSION = "0.12.0";
-export const SDK_VERSION = "^0.12.0";
+export const CLI_VERSION = "0.13.0";
+export const SDK_VERSION = "^0.13.0";
 export const STARTER_IDS = [
   "basic-readonly",
   "operational-action",